fix(data-sync-service): improve type handling and import structure in backtest routes and commodity signals
- Updated import statements in backtest_routes.py to enhance clarity and suppress type checking warnings.
- Modified return type in commodity_signals.py to ensure consistent float handling, improving type safety and preventing potential runtime issues.

+++ services/data-sync-service/src/data_sync_service/service/commodity_signals.py
@@ -53,7 +53,7 @@ def _rsi(closes: list[float], period: int = 14) -> float | None:
     if avg_loss == 0:
         return 100.0
     rs = avg_gain / avg_loss
-    return 100 - 100 / (1 + rs)
+    return float(100 - 100 / (1 + rs))
 
 def _ma(closes: list[float], n: int) -> float | None:
     if len(closes) < n:
